refactor(ui): split useDevicePrefs into composed preference hooks
